HighlightingEditor is format agnostic. Highlighting and 'applied actions' can be enabled independently.
Also 'applied actions' is called 'AutoFormat'
Moved IndentationFilter to MarkdownAutoFormat and added some methods to change/enable/disable highlighting.
