+++ app/src/main/java/net/gsantner/markor/format/highlighter/markdown/MarkdownAutoFormat.java
@@ -1,0 +1,131 @@
+package net.gsantner.markor.format.highlighter.markdown;
+
+import android.text.InputFilter;
+import android.text.Spanned;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * Created by stud on 21/11/17.
+ */
+
+public class MarkdownAutoFormat implements InputFilter {
+    private Pattern orderedListPrefixPattern = Pattern.compile("(?m)^([0-9]+)(\\.)");
+
+    @Override
+    public CharSequence filter(
+            CharSequence source,
+            int start,
+            int end,
+            Spanned dest,
+            int dstart,
+            int dend) {
+
+        if (true && //modified &&
+                end - start == 1 &&
+                start < source.length() &&
+                dstart <= dest.length()) {
+            char newChar = source.charAt(start);
+
+            if (newChar == '\n') {
+                return autoIndent(
+                        source,
+                        dest,
+                        dstart,
+                        dend);
+            }
+        }
+
+        return source;
+    }
+
+    private CharSequence autoIndent(CharSequence source, Spanned dest, int dstart, int dend) {
+
+        int istart = findLineBreakPosition(dest, dstart);
+
+        // append white space of previous line and new indent
+        return source + createIndentForNextLine(dest, dend, istart);
+    }
+
+    private int findLineBreakPosition(Spanned dest, int dstart) {
+        int istart = dstart - 1;
+
+        for (; istart > -1; --istart) {
+            char c = dest.charAt(istart);
+
+            if (c == '\n')
+                break;
+        }
+        return istart;
+    }
+
+    private String createIndentForNextLine(Spanned dest, int dend, int istart) {
+
+        if (istart > -1 && istart < dest.length() - 1) {
+            int iend;
+
+            for (iend = ++istart;
+                 iend < dest.length() - 1;
+                 ++iend) {
+                char c = dest.charAt(iend);
+
+                if (c != ' ' &&
+                        c != '\t') {
+                    break;
+                }
+            }
+
+            if (iend < dest.length() - 1) {
+                // This is for any line that is not the first line in a file
+                if (dest.charAt(iend + 1) == ' ') {
+                    return dest.subSequence(istart, iend) + addBulletPointIfNeeded(dest.charAt(iend));
+                } else {
+                    Matcher m = orderedListPrefixPattern.matcher(dest.toString().substring(iend));
+                    if (m.find()) {
+                        return dest.subSequence(istart, iend) + addNumericListItemIfNeeded(m.group(1));
+                    } else {
+                        return "";
+                    }
+                }
+            } else {
+                return "";
+            }
+        } else if (istart > -1) {
+            return "";
+        } else if (dest.length() > 1) {
+            // This is for the first line in a file.
+            if (dest.charAt(1) == ' ') {
+                return addBulletPointIfNeeded(dest.charAt(0));
+            } else {
+                Matcher m = orderedListPrefixPattern.matcher(dest.toString());
+                if (m.find()) {
+                    return addNumericListItemIfNeeded(m.group(1));
+                } else {
+                    return "";
+                }
+            }
+        } else {
+            return "";
+        }
+    }
+
+    private String addNumericListItemIfNeeded(String itemNumStr) {
+        try {
+            int nextC = Integer.parseInt(itemNumStr) + 1;
+            return nextC + ". ";
+        } catch (NumberFormatException e) {
+            // This should never ever happen
+            return "";
+        }
+    }
+
+    private String addBulletPointIfNeeded(char character) {
+        if (character == '*' || character == '+' || character == '-' || character == '>') {
+            return Character.toString(character) + " ";
+        } else {
+            return "";
+        }
+
+    }
+}
