+++ app/src/main/java/net/gsantner/markor/format/highlighter/markdown/MarkdownHighlighter.java
@@ -1,0 +1,17 @@
+package net.gsantner.markor.format.highlighter.markdown;
+
+import android.text.InputFilter;
+import android.text.Spanned;
+
+import net.gsantner.markor.format.highlighter.Highlighter;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * Created by stud on 21/11/17.
+ */
+
+public class MarkdownHighlighter{
+}
+
